Split chain type traits into separate modules
